requirements.txt
streamlit
numpy
google-generativeai
pandas
plotly